[VM, LotView] Various changes
- Support Head and Pelvis container slots.
- Add new TSO call modes and primitive modifiers for an upcoming event.
- Initial support for Manually Set Weather.
- TS1 improvements to CAS

+++ TSOClient/tso.simantics/Primitives/VMExpression.cs
@@ -132,7 +132,7 @@ namespace FSO.SimAntics.Engine.Primitives
                     lhsValue = VMMemory.GetBigVariable(context, operand.LhsOwner, operand.LhsData);
                     rhsValue = VMMemory.GetBigVariable(context, operand.RhsOwner, operand.RhsData);
                     var clearBitval = ~(1 << (rhsValue - 1));
-                    lhsValue &= (short)clearBitval;
+                    lhsValue &= clearBitval;
                     if (VMMemory.SetBigVariable(context, operand.LhsOwner, operand.LhsData, lhsValue))
                     {
                         return VMPrimitiveExitCode.GOTO_TRUE;
